lambda_functions/auth_lambda/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.0.1
